--- src/defines/env.ts
@@ -1,3 +1,0 @@
-const ENVIRONMENT_VARIABLES = ['JWT_SECRET', 'HASHIDS_KEY', 'MONGODB_URI', 'MONGODB_NAME'] as const;
-
-export type EnvironmentVariable = typeof ENVIRONMENT_VARIABLES[number];
